proyecto_final

+++ src/public/js/deleteProductAdmin.js
@@ -1,5 +1,5 @@
 const listProducts = document.getElementById("listProducts");
-const userId = document.getElementById("userId").textContent;
+const userId = document.querySelector(".uid").id;
 const card = document.getElementsByClassName("card");
 
 listProducts.addEventListener('click', (e) => {
@@ -7,14 +7,30 @@ listProducts.addEventListener('click', (e) => {
 })
 const buttonDelete = async(e) => {
     if(e.target.classList.contains('buttonDeleteProduct')){
-        // console.log(e.target.parentElement)
-        deleteProduct(e.target.parentElement)
+        const obj = e.target.parentElement;
+        let productId = obj.querySelector(".buttonDeleteProduct").id;
+        Swal.fire({
+            title: `Quieres eliminar este producto: ${productId}?`,
+            showDenyButton: true,
+            showCancelButton: false,
+            confirmButtonText: 'Yes',
+            denyButtonText: 'No',
+            customClass: {
+                actions: 'my-actions',
+                cancelButton: 'order-1 right-gap',
+                confirmButton: 'order-2',
+                denyButton: 'order-3',
+            }
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteProduct(e.target.parentElement)
+            }
+        })
     }
     e.stopPropagation();
 }
 
 const deleteProduct = async(obj) =>{
-    // console.log(obj,"producto")
     const productId = obj.querySelector(".buttonDeleteProduct").id;
 
     const response = await fetch(`api/products/${productId}`,{
@@ -25,5 +41,5 @@ const deleteProduct = async(obj) =>{
     })
     await response.json();
     swal.fire("Se elimino el producto")
-    window.location.replace('/products');
+    setTimeout(()=>window.location.reload(),2000);
 }
